+++ conda_index/index/__init__.py
@@ -732,7 +732,7 @@ class ChannelIndex:
         channeldata_file = os.path.join(self.output_root, "channeldata.json")
         return channeldata_file
 
-    def index_subdir(self, subdir, verbose=False, progress=False):
+    def index_subdir(self, subdir, verbose=False, progress=False) -> dict:
         """
         Return repodata from the cache without reading old repodata.json
 
@@ -743,7 +743,17 @@ class ChannelIndex:
 
         log.debug("Building repodata for %s/%s", self.channel_name, subdir)
 
-        new_repodata_packages, new_repodata_conda_packages = cache.indexed_packages()
+        # XXX add parameter to control record filter. We filter this way to
+        # avoid iterating over the packages another time, this may or may not be
+        # a useful optimization.
+        def filter(record):
+            return {
+                k: v
+                for k, v in record.items()
+                if not ((k == "subdir" and v == subdir) or k in ("platform", "arch"))
+            }
+
+        new_repodata_packages, new_repodata_conda_packages = cache.indexed_packages(filter=filter)
 
         new_repodata = {
             "packages": new_repodata_packages,